fix: add support for fedora and ubuntu

+++ bootstrap.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
 echo "Installing prereqs..."
-sudo apt install -y ansible build-essential python3
+if [[ -f $(which apt) ]]; then
+    sudo apt install -y ansible build-essential python3
+else
+    sudo dnf install -y ansible
+fi
 
 echo "Installing roles"
 ansible-galaxy install -r requirements.yml
